completed program, works as intended. need to work on makeup

+++ main.py
@@ -14,37 +14,59 @@ def addItem (list):
     NewItemType = input("Enter type of Item (Grocey, tobaco, or taxable): ")
     NewItemPrice = input("Enter the selling price: ")
     NewItemCost = input("Cost of the product: ")
-    newItemUnits = input("how many units being added: ")
+    newItemUnits = input("how many units being added: \n")
     list.append ( inventoryClass.item(NewItemName, NewItemExp, NewItemType,NewItemPrice, NewItemCost, newItemUnits) )
     
 #FIX ME need to understand how to delete an object based on its attribute, while looping through the list
 def removeItem(list):
+
+    if len(list) == 0:
+        print("Nothing to remove, no inventory to remove.\n")
+        return
     deleteItem = input("Enter the name of the item you want to remove: ")
     for i in range (len(list)):
         if list[i].name == deleteItem:
             del list[i]
 
+def searchItem(list, target):
+    
+    for i in range(len(list)):
+        if list[i].name == target:
+            print(list[i])
+            return
+
+    print("Item not found in inventory.\n")
     
 
-   
+def printInventory(list):
+    for obj in list:
+        print(obj)
 
-print("******************** Welcome to your Inventory Control System ********************")
+def switch(case, list, target):
+    if (case == 'A') or (case =='a'):
+        addItem(list)
+    elif (case == 'b') or (case == 'B'):
+        removeItem(list)
+    elif (case == 'c') or (case == 'C'):
+        searchItem(list, target)
+    elif (case == 'd') or (case == 'D'):
+        printInventory(list)
+    else: 
+        return
 
-print("What would you like to do? Choose one of the following: ")
-print("A: Add an item")
-#print("B: Remove an item")
-#print("C: search for an item inventory")
-#print("D: Print Inventory sheet")
 
+quitChar = 'o'
+newTarget = "searchItem"
 newList = []
-addItem(newList)
-#addItem(newList)
-for obj in newList:
-    print(obj)
 
-removeItem(newList)
+print("******************** Welcome to your Inventory Control System ********************\n")
+while not (quitChar == 'q' or  quitChar == 'Q'):
+   
+    print("What would you like to do? Choose one of the following: \n")
+    print("A: Add an item\n")
+    print("B: Remove an item\n")
+    print("C: search for an item inventory\n")
+    print("D: Print Inventory sheet\n")
 
-for obj in newList:
-    print(obj)
-
-
+    quitChar = input("\nchoose an option: ")
+    switch(quitChar, newList,newTarget)
